requirements.txt
streamlit>=1.28.0
pandas>=1.4.0
numpy>=1.23.0
httpx>=0.24.0
pillow>=9.0.0
python-dotenv>=1.0.0
matplotlib>=3.5.0
